Added test case and bugfix for sparsely filled columns.
This test demonstrates a bad assumption made in the code with the intent of fixing bad
behavior of old-IE. To support this, we use the value of `textContent` or `innerText`.

However, when the value of `textContent` is falsy (i.e. empty string) even on modern
browsers, the value of `innerText` is used. This works fine on Chrome and does the intended
thing on IE. However, Firefox returns `undefined` for this property. This causes the sort
to fail because comparisons with undefined are always false, thus all values are considered
equal.

The solution is to add a second fallback value that is the empty string, which was already
returned by `textContent`.